[feature] Posts can now include an updated date
Added an updated field on a post, to support modifications to the post
in the future. Wrote migrations for new field, and updated old
migrations to use the Django settings module.

Updated typing and documentation of the FlatPage loader
and library.

Updated all code to use Django settings module instead of directly
importing project settings.